+++ pkg/cmd/sessionLoginCmd.manual.go
@@ -50,8 +50,27 @@ func (n *sessionLoginCmd) initSession(cmd *cobra.Command, args []string) error {
 	if n.ClearExistingCookies {
 		client.SetCookies([]*http.Cookie{})
 	}
+
+	// If the password is not encrypted, then save it (which will apply the encryption)
+	if !cliConfig.IsPasswordEncrypted() {
+		Logger.Infof("Password is unencrypted. enforcing encryption")
+		n.onSave()
+	}
+
 	handler := c8ylogin.NewLoginHandler(client, cmd.ErrOrStderr(), n.onSave)
+
 	handler.TFACode = n.TFACode
 	handler.SetLogger(Logger)
-	return handler.Run()
+	err := handler.Run()
+
+	if err != nil {
+		return err
+	}
+
+	if handler.C8Yclient.TenantName != "" && cliConfig.GetTenant() != handler.C8Yclient.TenantName {
+		Logger.Infof("Saving tenant name")
+		n.onSave()
+	}
+
+	return nil
 }